requirements.txt
mysql-connector-python~=8.1.0
flask~=1.1.4
joblib~=1.3.2
pandas~=1.4.4
scikit-learn~=1.3.2
numpy~=1.21.6
scipy~=1.9.1